core code updates - enhancements
git-svn-id: http://svn.code.sf.net/p/hsqldb/svn/base/trunk@6089 7c7dc5f5-a22d-0410-a3af-b41755a11667

+++ src/org/hsqldb/persist/DataSpaceManagerBlocks.java
@@ -1098,7 +1098,7 @@ public class DataSpaceManagerBlocks implements DataSpaceManager {
      * output is in units of dataFileScale
      */
     long getPositionFromFileBlock(int fixedBlockPos) {
-        return fixedBlockPos * (fixedBlockSizeUnit / dataFileScale);
+        return fixedBlockPos * (long) (fixedBlockSizeUnit / dataFileScale);
     }
 
     int getFileBlockPosFromPosition(long position) {
